Allow shell-fm to build with --as-needed.
This patch splits the LDFLAGS variable, that should hold only the flags
used for linking, from LDLIBS, that in turn should be used to store the
libraries to link against. While the old behaviour is not an error, it
gives undefined references when compiling with --as-needed.

Reference Gentoo bug: https://bugs.gentoo.org/show_bug.cgi?id=286662